Fix edit-bill empty items + add numeric quantity migration
Two issues addressed:
1. When editing a bill whose items were never persisted (e.g. saved before the
   bill_items table existed), the status message now clearly tells the user to
   re-add medicines rather than showing a generic empty state.
2. Add alter-bill-items-quantity.sql migration to change bill_items.quantity
   from INTEGER to NUMERIC(10,3) so fractional quantities (e.g. 0.33 for 5 of
   15 tablets) can be stored. Also updated add-billing-tables.sql baseline.

+++ add-billing-tables.sql
@@ -26,7 +26,7 @@ create table if not exists bill_items (
   medicine_id    uuid,
   medicine_name  text          not null,
   location       text          not null default '',
-  quantity       integer       not null default 1,
+  quantity       numeric(10,3) not null default 1,
   mrp            numeric(10,2),
   purchase_price numeric(10,2),
   sell_price     numeric(10,2) not null,
